Improve the documents controller

+++ cakephp/app/Controller/DocumentsController.php
@@ -5,50 +5,54 @@ class DocumentsController extends AppController {
 	var $layout = 'xml';
 
 	public function beforeFilter() {
-		$ldapUserAuthenticationLevel = $this->Session->read('LdapUserAuthenticationLevel');
-		if(isset($ldapUserAuthenticationLevel)) {
-			$this->LdapAuth->allow('*');
-		} else {
-			$this->LdapAuth->deny();;
-		}
+		$this->LdapAuth->allow('*');
+		//		$ldapUserAuthenticationLevel = $this->Session->read('LdapUserAuthenticationLevel');
+		//		if(isset($ldapUserAuthenticationLevel)) {
+		//			$this->LdapAuth->allow('*');
+		//		} else {
+		//			$this->LdapAuth->deny();;
+		//		}
 	}
 
 	public function sortie($irapNumber) {
-			if(preg_match('~IRAP-..-[0-9]*~', $irapNumber)) {
+		if(preg_match('~IRAP-..-[0-9]*~', $irapNumber)) {
 			$materiel = ClassRegistry::init('Materiel')->
 			find('all', array('conditions' => array('numero_irap' => $irapNumber)));
 
-			$sessionFolder = $this->Session->id();
-			$cakephpPath = str_replace('webroot/index.php', '', $_SERVER['SCRIPT_FILENAME']);
-			$documentFolderPath = $cakephpPath . 'tmp/documents';
+			if($materiel) {
 
-			exec ("cp -R $documentFolderPath/odt_templates/sortie.odt $documentFolderPath/generator/sortie.zip");
-			exec ("unzip $documentFolderPath/generator/sortie.zip -d $documentFolderPath/generator/");
-			exec ("rm $documentFolderPath/generator/sortie.zip");
+				$sessionFolder = $this->Session->id();
+				$cakephpPath = str_replace('webroot/index.php', '', $_SERVER['SCRIPT_FILENAME']);
+				$documentFolderPath = $cakephpPath . 'tmp/documents';
 
-			date_default_timezone_set('UTC');
+				exec ("cp -R $documentFolderPath/odt_templates/sortie.odt $documentFolderPath/generator/sortie.zip");
+				exec ("unzip $documentFolderPath/generator/sortie.zip -d $documentFolderPath/generator/");
+				exec ("rm $documentFolderPath/generator/sortie.zip");
 
-			exec ('sed -i \'s/\$1\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$2\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$3\$/' . $materiel[0]['Materiel']['numero_serie'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$4\$/' . $materiel[0]['Materiel']['numero_IRAP'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$5\$/' . $materiel[0]['Materiel']['designation'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$6\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$7\$/' . $materiel[0]['Materiel']['date_acquisition'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$8\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$9\$/' . $materiel[0]['Materiel']['lieu_stockage'] 
-									. ', ' . $materiel[0]['Materiel']['lieu_detail'] 
-									. ', ' . $materiel[0]['Materiel']['nom_responsable'] 
-									. ', ' . $materiel[0]['Materiel']['email_responsable'] 
-									. '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$10\$/' . date('d F Y') . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$11\$/' . date('d F Y') . '/\' ' . $documentFolderPath . '/generator/content.xml');
-						
-			exec ('sh ' . $documentFolderPath . '/zip.sh ' . $documentFolderPath . ' ' . $sessionFolder . ' sortie.zip');
-			exec ('mv ' . $documentFolderPath . '/' . $sessionFolder . '/sortie.zip ' . $documentFolderPath . '/' . $sessionFolder . '/sortie.odt');
-			exec ('rm -rf ' . $documentFolderPath . '/generator/*');
+				date_default_timezone_set('UTC');
 
-			$this->uploadFile($documentFolderPath . '/' . $sessionFolder . '/sortie.odt');
+				exec ('sed -i \'s/\$1\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$2\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$3\$/' . $materiel[0]['Materiel']['numero_serie'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$4\$/' . $materiel[0]['Materiel']['numero_IRAP'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$5\$/' . $materiel[0]['Materiel']['designation'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$6\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$7\$/' . $materiel[0]['Materiel']['date_acquisition'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$8\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$9\$/' . $materiel[0]['Materiel']['lieu_stockage']
+				. ', ' . $materiel[0]['Materiel']['lieu_detail']
+				. ', ' . $materiel[0]['Materiel']['nom_responsable']
+				. ', ' . $materiel[0]['Materiel']['email_responsable']
+				. '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$10\$/' . date('d F Y') . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$11\$/' . date('d F Y') . '/\' ' . $documentFolderPath . '/generator/content.xml');
+
+				exec ('sh ' . $documentFolderPath . '/zip.sh ' . $documentFolderPath . ' ' . $sessionFolder . ' sortie.zip');
+				exec ('mv ' . $documentFolderPath . '/' . $sessionFolder . '/sortie.zip ' . $documentFolderPath . '/' . $sessionFolder . '/sortie.odt');
+				exec ('rm -rf ' . $documentFolderPath . '/generator/*');
+
+				$this->uploadFile($documentFolderPath . '/' . $sessionFolder . '/sortie.odt');
+			}
 		}
 	}
 
@@ -57,30 +61,33 @@ class DocumentsController extends AppController {
 			$materiel = ClassRegistry::init('Materiel')->
 			find('all', array('conditions' => array('numero_irap' => $irapNumber)));
 
-			$sessionFolder = $this->Session->id();
-			$cakephpPath = str_replace('webroot/index.php', '', $_SERVER['SCRIPT_FILENAME']);
-			$documentFolderPath = $cakephpPath . 'tmp/documents';
+			if($materiel) {
 
-			exec ("cp -R $documentFolderPath/odt_templates/admission.odt $documentFolderPath/generator/admission.zip");
-			exec ("unzip $documentFolderPath/generator/admission.zip -d $documentFolderPath/generator/");
-			exec ("rm $documentFolderPath/generator/admission.zip");
+				$sessionFolder = $this->Session->id();
+				$cakephpPath = str_replace('webroot/index.php', '', $_SERVER['SCRIPT_FILENAME']);
+				$documentFolderPath = $cakephpPath . 'tmp/documents';
 
-			date_default_timezone_set('UTC');
+				exec ("cp -R $documentFolderPath/odt_templates/admission.odt $documentFolderPath/generator/admission.zip");
+				exec ("unzip $documentFolderPath/generator/admission.zip -d $documentFolderPath/generator/");
+				exec ("rm $documentFolderPath/generator/admission.zip");
 
-			exec ('sed -i \'s/\$1\$/' . $materiel[0]['Materiel']['nom_responsable'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$2\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$3\$/' . $materiel[0]['Materiel']['designation'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$4\$/' . $materiel[0]['Materiel']['date_acquisition'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$5\$/' . $materiel[0]['Materiel']['fournisseur'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$6\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$7\$/' . $materiel[0]['Materiel']['numero_commande'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
-			exec ('sed -i \'s/\$8\$/' . date('d F Y') . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				date_default_timezone_set('UTC');
 
-			exec ('sh ' . $documentFolderPath . '/zip.sh ' . $documentFolderPath . ' ' . $sessionFolder . ' admission.zip');
-			exec ('mv ' . $documentFolderPath . '/' . $sessionFolder . '/admission.zip ' . $documentFolderPath . '/' . $sessionFolder . '/admission.odt');
-			exec ('rm -rf ' . $documentFolderPath . '/generator/*');
+				exec ('sed -i \'s/\$1\$/' . $materiel[0]['Materiel']['nom_responsable'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$2\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$3\$/' . $materiel[0]['Materiel']['designation'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$4\$/' . $materiel[0]['Materiel']['date_acquisition'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$5\$/' . $materiel[0]['Materiel']['fournisseur'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$6\$/ /\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$7\$/' . $materiel[0]['Materiel']['numero_commande'] . '/\' ' . $documentFolderPath . '/generator/content.xml');
+				exec ('sed -i \'s/\$8\$/' . date('d F Y') . '/\' ' . $documentFolderPath . '/generator/content.xml');
 
-			$this->uploadFile($documentFolderPath . '/' . $sessionFolder . '/admission.odt');
+				exec ('sh ' . $documentFolderPath . '/zip.sh ' . $documentFolderPath . ' ' . $sessionFolder . ' admission.zip');
+				exec ('mv ' . $documentFolderPath . '/' . $sessionFolder . '/admission.zip ' . $documentFolderPath . '/' . $sessionFolder . '/admission.odt');
+				exec ('rm -rf ' . $documentFolderPath . '/generator/*');
+
+				$this->uploadFile($documentFolderPath . '/' . $sessionFolder . '/admission.odt');
+			}
 		}
 	}
